renamed the main package to zeeguu_teacher_dashboard
