+++ zeeguu_dashboard/zeeguu_teacher_dashboard/__init__.py
@@ -3,4 +3,4 @@ from flask import Flask
 app = Flask(__name__)
 
 # These imports have be issued after the app is initialized due to how Flask works.
-from app.page_routes import homepage, errorpages, login, studentpage, classroom
+from .page_routes import homepage, errorpages, login, studentpage, classroom
