+++ zeeguu_dashboard/zeeguu_teacher_dashboard/static/css/style.css
