+++ zeeguu_dashboard/zeeguu_teacher_dashboard/templates/edit_class.html
