+++ zeeguu_dashboard/zeeguu_teacher_dashboard/util/permissions.py
@@ -1,7 +1,7 @@
 from functools import wraps
 from flask import redirect, session
-from app.api.api_connection import api_get
-from app import app
+from zeeguu_teacher_dashboard.api.api_connection import api_get
+from zeeguu_teacher_dashboard import app
 
 """
 This file contains all of the functions responsible for validating the user on different page_routes.
