+++ zeeguu_dashboard/zeeguu_teacher_dashboard/util/user.py
@@ -1,6 +1,6 @@
 import json
 
-from app.api.api_connection import api_get
+from zeeguu_teacher_dashboard.api.api_connection import api_get
 
 """
 This file contains all of the utility functions for loading and formatting user data.
